refactor to create abstract base class for 2 checkers
