+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-azure-sdk/src/main/resources/META-INF/azure-intellij-plugin-azure-sdk.xml
@@ -138,16 +138,32 @@
 
         <localInspection
                 language="JAVA"
-                shortName="UpdateCheckpointAsyncCheck"
-                displayName="Update Checkpoint Async Check"
+                shortName="UpdateCheckpointAsyncSubscribeChecker"
+                displayName="Update Checkpoint Async Subscribe Checker"
                 groupName="Azure SDK Rules"
                 enabledByDefault="true"
                 level="WARNING"
-                implementationClass="com.microsoft.azure.toolkit.intellij.azure.sdk.buildtool.UpdateCheckpointAsyncCheck">
+                implementationClass="com.microsoft.azure.toolkit.intellij.azure.sdk.buildtool.UpdateCheckpointAsyncSubscribeChecker">
 
             <class>
-                com.microsoft.azure.toolkit.intellij.azure.sdk.buildtool.UpdateCheckpointAsyncCheck
+                com.microsoft.azure.toolkit.intellij.azure.sdk.buildtool.UpdateCheckpointAsyncSubscribeChecker
             </class>
         </localInspection>
+
+        <localInspection
+                language="JAVA"
+                shortName="UpdateCheckpointAsyncBlockChecker"
+                displayName="Update Checkpoint Async Block Checker"
+                groupName="Azure SDK Rules"
+                enabledByDefault="true"
+                level="WARNING"
+                implementationClass="com.microsoft.azure.toolkit.intellij.azure.sdk.buildtool.UpdateCheckpointAsyncBlockChecker">
+
+            <class>
+                com.microsoft.azure.toolkit.intellij.azure.sdk.buildtool.UpdateCheckpointAsyncBlockChecker
+            </class>
+        </localInspection>
+
+
     </extensions>
 </idea-plugin>
